+++ player.go
@@ -6,24 +6,36 @@ import (
 )
 
 const (
-	playerObjectPath        = "/org/mpris/MediaPlayer2"
-	playerInterface         = "org.mpris.MediaPlayer2.Player"
-	playerNextMethod        = playerInterface + ".Next"
-	playerPreviousMethod    = playerInterface + ".Previous"
-	playerPauseMethod       = playerInterface + ".Pause"
-	playerPlayPauseMethod   = playerInterface + ".PlayPause"
-	playerStopMethod        = playerInterface + ".Stop"
-	playerPlayMethod        = playerInterface + ".Play"
-	playerSeekMethod        = playerInterface + ".SeekTo"
-	playerSetPositionMethod = playerInterface + ".SetPosition"
-	playerOpenURIMethod     = playerInterface + ".OpenURI"
+	playerObjectPath             = "/org/mpris/MediaPlayer2"
+	playerInterface              = "org.mpris.MediaPlayer2.Player"
+	playerNextMethod             = playerInterface + ".Next"
+	playerPreviousMethod         = playerInterface + ".Previous"
+	playerPauseMethod            = playerInterface + ".Pause"
+	playerPlayPauseMethod        = playerInterface + ".PlayPause"
+	playerStopMethod             = playerInterface + ".Stop"
+	playerPlayMethod             = playerInterface + ".Play"
+	playerSeekMethod             = playerInterface + ".SeekTo"
+	playerSetPositionMethod      = playerInterface + ".SetPosition"
+	playerOpenURIMethod          = playerInterface + ".OpenURI"
+	playerPlaybackStatusProperty = playerInterface + ".PlaybackStatus"
 )
 
 var dbusSessionBus = dbus.SessionBus
 
-//go:generate moq -out dbus-wrapper_moq_test.go . dbusWrapperI
-type dbusWrapperI interface {
-	Call(dest string, path dbus.ObjectPath, method string, flags dbus.Flags, args ...interface{})
+//go:generate moq -out dbus-conn_moq_test.go . dbusConn
+type dbusConn interface {
+	Object(dest string, path dbus.ObjectPath) dbusBusObject
+}
+
+//go:generate moq -out dbus-bus-object_moq_test.go . dbusBusObject
+type dbusBusObject interface {
+	Call(method string, flags dbus.Flags, args ...interface{}) dbusCall
+	GetProperty(p string) (v dbus.Variant, e error)
+}
+
+//go:generate moq -out dbus-call_moq_test.go . dbusCall
+type dbusCall interface {
+	Store(retvalues ...interface{}) error
 }
 
 //Player is a implementation of dbus org.mpris.MediaPlayer2.Player. see: https://specifications.freedesktop.org/mpris-spec/2.2/Player_Interface.html
@@ -31,7 +43,7 @@ type dbusWrapperI interface {
 //Use NewPlayerWithConnection when you want to use a self-configured dbus.Conn
 type Player struct {
 	name       string
-	connection dbusWrapperI
+	connection dbusConn
 }
 
 //NewPlayer returns a new Player which is already connected to session-bus via dbus.SessionBus.
@@ -43,18 +55,20 @@ func NewPlayer(name string) (Player, error) {
 
 	return Player{
 		name: name,
-		connection: dbusWrapper{
+		connection: &dbusConnWrapper{
 			conn: connection,
 		},
 	}, nil
 }
 
 //NewPlayer returns a new Player with the given name and connection.
-func NewPlayerWithConnection(name string, connection *dbus.Conn) (Player, error) {
+func NewPlayerWithConnection(name string, connection *dbus.Conn) Player {
 	return Player{
-		name:       name,
-		connection: dbusWrapper{conn: connection},
-	}, nil
+		name: name,
+		connection: &dbusConnWrapper{
+			conn: connection,
+		},
+	}
 }
 
 //Next skips to the next track in the tracklist.
@@ -63,7 +77,7 @@ func NewPlayerWithConnection(name string, connection *dbus.Conn) (Player, error)
 //If CanGoNext is false, attempting to call this method should have no effect.
 //see: https://specifications.freedesktop.org/mpris-spec/latest/Player_Interface.html#Method:Next
 func (p Player) Next() {
-	p.connection.Call(p.name, playerObjectPath, playerNextMethod, 0)
+	p.connection.Object(p.name, playerObjectPath).Call(playerNextMethod, 0)
 }
 
 //Previous skips to the previous track in the tracklist.
@@ -72,7 +86,7 @@ func (p Player) Next() {
 //If CanGoPrevious is false, attempting to call this method should have no effect.
 //see: https://specifications.freedesktop.org/mpris-spec/latest/Player_Interface.html#Method:Previous
 func (p Player) Previous() {
-	p.connection.Call(p.name, playerObjectPath, playerPreviousMethod, 0)
+	p.connection.Object(p.name, playerObjectPath).Call(playerPreviousMethod, 0)
 }
 
 //Pause pauses playback.
@@ -81,7 +95,7 @@ func (p Player) Previous() {
 //If CanPause is false, attempting to call this method should have no effect.
 //see: https://specifications.freedesktop.org/mpris-spec/latest/Player_Interface.html#Method:Pause
 func (p Player) Pause() {
-	p.connection.Call(p.name, playerObjectPath, playerPauseMethod, 0)
+	p.connection.Object(p.name, playerObjectPath).Call(playerPauseMethod, 0)
 }
 
 //Play starts or resumes playback.
@@ -91,7 +105,7 @@ func (p Player) Pause() {
 //If CanPlay is false, attempting to call this method should have no effect.
 //see: https://specifications.freedesktop.org/mpris-spec/latest/Player_Interface.html#Method:Play
 func (p Player) Play() {
-	p.connection.Call(p.name, playerObjectPath, playerPlayMethod, 0)
+	p.connection.Object(p.name, playerObjectPath).Call(playerPlayMethod, 0)
 }
 
 //PlayPause pauses playback.
@@ -100,7 +114,7 @@ func (p Player) Play() {
 //If CanPause is false, attempting to call this method should have no effect and raise an error.
 //see: https://specifications.freedesktop.org/mpris-spec/latest/Player_Interface.html#Method:PlayPause
 func (p Player) PlayPause() {
-	p.connection.Call(p.name, playerObjectPath, playerPlayPauseMethod, 0)
+	p.connection.Object(p.name, playerObjectPath).Call(playerPlayPauseMethod, 0)
 }
 
 //Stop stops playback.
@@ -109,7 +123,7 @@ func (p Player) PlayPause() {
 //If CanControl is false, attempting to call this method should have no effect and raise an error.
 //see: https://specifications.freedesktop.org/mpris-spec/latest/Player_Interface.html#Method:Stop
 func (p Player) Stop() {
-	p.connection.Call(p.name, playerObjectPath, playerStopMethod, 0)
+	p.connection.Object(p.name, playerObjectPath).Call(playerStopMethod, 0)
 }
 
 //SeekTo seeks forward in the current track by the specified number of microseconds.
@@ -120,7 +134,7 @@ func (p Player) Stop() {
 //If the CanSeek property is false, this has no effect.
 //see: https://specifications.freedesktop.org/mpris-spec/latest/Player_Interface.html#Method:Seek
 func (p Player) SeekTo(offset int64) {
-	p.connection.Call(p.name, playerObjectPath, playerSeekMethod, 0, offset)
+	p.connection.Object(p.name, playerObjectPath).Call(playerSeekMethod, 0, offset)
 }
 
 //SetPosition Sets the current track position in microseconds.
@@ -133,7 +147,7 @@ func (p Player) SeekTo(offset int64) {
 //If the CanSeek property is false, this has no effect.
 //see: https://specifications.freedesktop.org/mpris-spec/latest/Player_Interface.html#Method:SetPosition
 func (p Player) SetPosition(trackID dbus.ObjectPath, position int64) {
-	p.connection.Call(p.name, playerObjectPath, playerSetPositionMethod, 0, trackID, position)
+	p.connection.Object(p.name, playerObjectPath).Call(playerSetPositionMethod, 0, trackID, position)
 }
 
 //OpenURI opens the Uri given as an argument
@@ -145,5 +159,17 @@ func (p Player) SetPosition(trackID dbus.ObjectPath, position int64) {
 //If the media player implements the TrackList interface, then the opened track should be made part of the tracklist, the org.mpris.MediaPlayer2.TrackList.TrackAdded or org.mpris.MediaPlayer2.TrackList.TrackListReplaced signal should be fired, as well as the org.freedesktop.DBus.Properties.PropertiesChanged signal on the tracklist interface.
 //see: https://specifications.freedesktop.org/mpris-spec/latest/Player_Interface.html#Method:OpenUri
 func (p Player) OpenURI(uri string) {
-	p.connection.Call(p.name, playerObjectPath, playerOpenURIMethod, 0, uri)
+	p.connection.Object(p.name, playerObjectPath).Call(playerOpenURIMethod, 0, uri)
+}
+
+//PlaybackStatus returns the current playback status.
+//When this property changes, the org.freedesktop.DBus.Properties.PropertiesChanged signal is emitted with the new value.
+//May be "Playing", "Paused" or "Stopped".
+func (p Player) PlaybackStatus() (string, error) {
+	v, err := p.connection.Object(p.name, playerObjectPath).GetProperty(playerPlaybackStatusProperty)
+	if err != nil {
+		return "", fmt.Errorf("failed to get property %q: %w", playerPlaybackStatusProperty, err)
+	}
+
+	return v.Value().(string), nil
 }