+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 ## unreleased
-- add mpris MediaPlayer2.Player methods
+- add mpris MediaPlayer2.Player methods (all)
+- add mpris MediaPlayer2.Player properties (PlaybackStatus)